// front/src/app/enum/phase.enum.ts
export enum SelectablePhaseEnum {
    undefined = 'UNDEFINED',
    development = 'DEVELOPMENT',
    construction = 'CONSTRUCTION',
    action = 'ACTION',
    production = 'PRODUCTION',
    research = 'RESEARCH'
}
export enum NonSelectablePhaseEnum {
    undefined = 'UNDEFINED',
    planification = 'PLANIFICATION',
    development = 'DEVELOPMENT',
    construction = 'CONSTRUCTION',
    action = 'ACTION',
    production = 'PRODUCTION',
    research = 'RESEARCH'
}